refactor: update validation middleware across multiple routes
- Replaced instances of the old validation middleware with the new enhanced-validation middleware in various route files, improving request validation consistency.
- Removed the obsolete migrate-payment-intent-index script as it is no longer needed.
- Enhanced media service with Redis health checks for better error handling during cache operations.
