+++ backend/routes/cart.route.ts
@@ -1,7 +1,7 @@
 import { Router } from 'express';
 import { getCartProducts, addToCart, removeFromCart, updateQuantity } from '../controllers/cart.controller.js';
 import { protectRoute } from '../middleware/auth.middleware.js';
-import { validateBody, validateParams } from '../middleware/validation.middleware.js';
+import { validateBody, validateParams } from '../middleware/enhanced-validation.middleware.js';
 import { validateAddToCart, validateCartInventory, includeCartValidationWarnings } from '../middleware/cart-validation.middleware.js';
 import { addToCartSchema, updateQuantitySchema, productIdParamSchema, cartProductIdParamSchema } from '../validations/index.js';
 import { asyncHandler } from '../utils/asyncHandler.js';
